feat: Dashboard revisited
Dashboard completely overhauled to provide more streamlined and simpler
UX. Recent repositories and categorised repositories
collections are combined. The repositories list is now sync'ed with the
list of repositories dropdown to provide the consistent UX.
The list is managed via the FormRecentReposSettings.

The dashboard can have two distinct themes - light and dark.
A theme can be selected via Settings > Colors tab. If no theme is
selected then the theme is automatically chosen based on the current
Windows theme.

Also:
- DashboardCategory deprecated and removed
- DashboardItem deprecated and removed
- DashboardEditor deprecated and removed
- Repository category types removed
- RSS feeds deprecated and removed

#3658