Alternative architecture proposal
1. Load custom diff/merge tools in a dedicated CustomDiffMergeToolProvider
It also responsible for caching, which is written in a rather naive way and
based on the current working directory. If the working directory is changed
the cache is busted.
I don't think we can expect race conditions from multiple simultaneous
threads working in different repos. E.g. submodules are opened in new
app instances.

2. UI layer will get the list of tools from CustomDiffMergeToolProvider and
bind to the UI.
These routines can likely be generalised and unified.